requirements.txt
python-dotenv==1.*
tweepy==4.*
Jinja2==3.*
polygon-api-client==1.*
google-generativeai==0.7.*
requests==2.*
openai>=1.40

